+++ Changes.md
@@ -9,3 +9,4 @@ Version 3.3.0.1
   3. Added 'nediconsleft' option.
   4. Added 'sectioncontentjustification' option.  Note: Uses 'Flexbox' so might not be the solution if old browsers used.
   5. Added activity tracking background colours.
+  6. Added all 'showsection0' options, being 'hide', 'show' and 'show only section 0'.